Improvements to 64-bit support to ensure data type consistency.

+++ LookIO/Source/Controllers/LIOAltChatViewController.h
@@ -99,7 +99,7 @@
     int currentPopoverType;
 
     LIOChatMessage *clickedFailedMessage;
-    int clickedFailedMessageIndex;
+    NSInteger clickedFailedMessageIndex;
     
     BOOL isAnimatingDismissal;
