+++ LookIO/Source/Other/Survey/LIOSurveyQuestion.h
@@ -27,11 +27,11 @@ typedef enum
 
 @interface LIOSurveyQuestion : NSObject
 {
-    int questionId;
+    NSInteger questionId;
     BOOL mandatory;
-    int order;
+    NSInteger order;
     NSString *label;
-    int logicId;
+    NSInteger logicId;
     LIOSurveyQuestionDisplayType displayType;
     LIOSurveyQuestionValidationType validationType;
     NSArray *pickerEntries;
@@ -39,11 +39,11 @@ typedef enum
     NSString* lastKnownValue;
 }
 
-@property(nonatomic, assign) int questionId;
+@property(nonatomic, assign) NSInteger questionId;
 @property(nonatomic, assign) BOOL mandatory;
-@property(nonatomic, assign) int order;
+@property(nonatomic, assign) NSInteger order;
 @property(nonatomic, retain) NSString *label;
-@property(nonatomic, assign) int logicId;
+@property(nonatomic, assign) NSInteger logicId;
 @property(nonatomic, assign) LIOSurveyQuestionDisplayType displayType;
 @property(nonatomic, assign) LIOSurveyQuestionValidationType validationType;
 @property(nonatomic, retain) NSArray *pickerEntries;
